+++ Users/Super/User/ajax_grid_data.php
@@ -53,7 +53,7 @@ while( $row=mysqli_fetch_array($query) ) {  // preparing an array
     $nestedData[] = $row["tipo"];
     $nestedData[] = '<td><center>
                      <a href="updateUser.php?id='.$row['id_usu'].'"  data-toggle="tooltip" title="Editar datos" class="btn btn-sm btn-info"> <i class="fa fa-fw fa-pencil-alt"></i> </a>
-                     <a href="showUser.php?action=delete&id='.$row['id_usu'].'"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa fa-fw fa-trash"></i> </a>
+                     <a href="showUser.php?action=delete&id='.$row['id_usu'].'"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger" onclick="return confirm(\'Estas seguro de elimar el usuario?\');"> <i class="fa fa-fw fa-trash"></i> </a>
 				     </center></td>';
 
     $data[] = $nestedData;
